requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
aiofiles==23.2.1
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.4.2
httpx==0.25.2